@@ -1,6 +1,6 @@
 package com.group.libraryapp.domain.book;
 
-import javax.persistence.*;
+import jakarta.persistence.*;
 
 @Entity
 public class Book {
