@@ -45,9 +45,10 @@ public class UserServiceV2 {
 
     @Transactional
     public void deleteUser(String name) {
+        System.out.println("1");
         User user = userRepository.findByName(name)
-                .orElseThrow(IllegalAccessError::new);
-
+                .orElseThrow(IllegalArgumentException::new);
+        System.out.println("2");
         userRepository.delete(user);
     }
 }
